Organize integration test fixtures / remove extra fixture modules
