+++ cstar/tests/conftest.py
@@ -1,10 +1,23 @@
+import typing as t
+from pathlib import Path
 from unittest.mock import Mock, patch
 
 import pytest
 
 
+@pytest.fixture(scope="session")
+def tests_path() -> Path:
+    """Fixture that returns the path to the directory containing C-Star tests.
+
+    Returns
+    -------
+    Path
+    """
+    return Path(__file__).parent
+
+
 @pytest.fixture(scope="session", autouse=True)
-def skip_html_checks_for_performance():
+def skip_html_checks_for_performance() -> t.Generator[None]:
     """
     Global auto-use that skips the head checks used to see if HTTP input datasets are HTML.
     For our test cases, they are always files, except for when we explicitly test that they,
@@ -28,3 +41,38 @@ def mock_user_env_name() -> str:
         The name of the .env file
     """
     return ".mock.env"
+
+
+@pytest.fixture
+def mock_lmod_filename() -> str:
+    """Provide a unique .lmod filename for tests.
+
+    Returns
+    -------
+    str
+        The filename
+    """
+    return "mock.lmod"
+
+
+@pytest.fixture
+def mock_lmod_path(tmp_path: Path, mock_lmod_filename: str) -> Path:
+    """Create an empty, temporary .lmod file and return the path.
+
+    Parameters
+    ----------
+    tmp_path : Path
+        The path to a temporary location to write the lmod file
+    mock_lmod_filename : str
+        The filename to use for the .lmod file
+
+    Returns
+    -------
+    str
+        The complete path to the file
+    """
+    tmp_path.mkdir(parents=True, exist_ok=True)
+
+    path = tmp_path / mock_lmod_filename
+    path.touch()  # CStarEnvironment expects the file to exist & opens it
+    return path
